use default region from aws config, closes #3

+++ internal/cmd/delete.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/pete911/aws-vpn/internal/cmd/prompt"
 	"github.com/spf13/cobra"
 	"os"
 )
@@ -27,8 +28,11 @@ func runDelete(cmd *cobra.Command, args []string) {
 
 	logger := NewLogger()
 	client := NewClient(logger)
-
 	instance := SelectInstance(client, name)
+	if !prompt.Prompt(fmt.Sprintf("delete %s VPN instance in %s region", instance, client.Region)) {
+		return
+	}
+
 	if err := client.Delete(instance); err != nil {
 		fmt.Printf("delete %s VPN: %v\n", instance.Name, err)
 		os.Exit(1)
